JDBC - Organize MSSQL tests into MSSQLDriverTest file
